test: added tests for Swagger/OpenAPI import via file and URL
