+++ gravitee-apim-cypress/cypress/fixtures/json/petstore_swaggerv2.json
@@ -13,7 +13,7 @@
       "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
     }
   },
-  "host": "petstore:8080",
+  "host": "petstore.swagger.io",
   "basePath": "/v2",
   "tags": [
     {
