+++ gravitee-apim-cypress/package.json
@@ -1,6 +1,6 @@
 {
     "name": "gravitee-apim-cypress",
-    "version": "3.13.0-SNAPSHOT",
+    "version": "3.16.1",
     "private": true,
     "description": "Gravitee.io APIM - Cypress",
     "scripts": {
@@ -25,7 +25,6 @@
     "devDependencies": {
         "@types/faker": "5.5.9",
         "@types/node": "16.10.9",
-        "@types/semver": "^7.3.9",
         "ansi-regex": "6.0.1",
         "cypress": "9.5.0",
         "cypress-terminal-report": "3.4.1",
@@ -33,7 +32,6 @@
         "jsonwebtoken": "8.5.1",
         "license-check-and-add": "4.0.3",
         "prettier": "2.5.1",
-        "semver": "^7.3.5",
         "typescript": "4.5.2"
     },
     "dependencies": {
